requirements.txt
transformers
timm
scikit-learn
Pillow
requests
numpy
rasterio
shapely
pyproj
boto3
geopy